Introduce `WebAuthnUserHandle` to persist a truly random user handle
Replaces the previous approach of hashing the users' primary key and
implements the guidance in https://www.w3.org/TR/webauthn-3/#sctn-user-handle-privacy

This yields better privacy, as there is no way now to link a user
handle back to a user account. See the security note about a possible
de-anonymization attack vector removed by this commit.

When I originally wrote the handle generation code and said security
note, I was under the impression that storing random bytes would have to
be done on the user model. Requiring developers to implement a
`WebAuthnUserHandle` mixin or the like in their custom user model to add
this field. That would be more steps and add unwanted bloat.

With the OneToOneField approach it works more transparently, with no
effort required from the developer. Wish I thought of this from the
beginning.