Mark `test_checks` testcases as requiring database access
`manage.py check` now accesses the database, this change is
scheduled to be released in Django 6.1 and this new behavior
surfaced as test failures in this library's nightly tests against
Django's `main` branch.

Bug report was raised with Django (#36739) but considered invalid, since
`manage.py check` may interact with the database or filesystem. This is
considered to be expected behavior by the Django developers. It just
did not happen to do so before.

Easiest fix for us is to mark our test cases as needing database access.

References:

- https://code.djangoproject.com/ticket/36739